lots of things that annoyed me
