+++ randomizer/CompileHints.py
@@ -589,23 +589,36 @@ def compileHints(spoiler: Spoiler) -> bool:
             if spoiler.settings.key_8_helm:  # You may know that Key 8 is in Helm and that's pointless to hint
                 item_region_locations_to_hint.remove(Locations.HelmKey)
         # Determine what moves are hintable
-        all_hintable_moves = ItemPool.AllKongMoves() + ItemPool.TrainingBarrelAbilities()
+        all_hintable_moves = ItemPool.AllKongMoves() + ItemPool.TrainingBarrelAbilities() + kongs_to_hint
         if spoiler.settings.shockwave_status != ShockwaveStatus.start_with:
             all_hintable_moves.extend(ItemPool.ShockwaveTypeItems(spoiler.settings))
         if spoiler.settings.shuffle_items and Types.Bean in spoiler.settings.shuffled_location_types:
             all_hintable_moves.append(Items.Bean)
         optional_hintable_locations = []
+        slam_locations = []
         # Loop through all locations, finding the location of all of these hintable moves
         for id, location in spoiler.LocationList.items():
+            # Note the location of slams - these will always be at least optionally hintable and sometimes required to be hinted
+            if location.item == Items.ProgressiveSlam:
+                slam_locations.append(id)
+            # Never hint training moves for obvious reasons
+            if location.type in (Types.TrainingBarrel, Types.PreGivenMove):
+                continue
             # If it's a woth item, it must be hinted so put it in the list
-            if id in spoiler.woth_locations and location.type not in (Types.TrainingBarrel, Types.PreGivenMove):
+            if id in spoiler.woth_locations:
                 if location.item in kongs_to_hint:
                     item_region_locations_to_hint.insert(0, id)
                 elif location.item in all_hintable_moves:
                     item_region_locations_to_hint.append(id)
             # To be hintable, it can't be a starting move
-            elif location.item in all_hintable_moves and location.type not in (Types.TrainingBarrel, Types.PreGivenMove):
+            elif location.item in all_hintable_moves:
                 optional_hintable_locations.append(id)
+        # If there's room, always hint a slam if we haven't hinted one already
+        hinted_slam_locations = [loc for loc in slam_locations if loc in item_region_locations_to_hint or spoiler.LocationList[loc].type in (Types.TrainingBarrel, Types.PreGivenMove)]
+        if len(item_region_locations_to_hint) < hint_distribution[HintType.ItemRegion] and len(hinted_slam_locations) < 2:
+            loc_to_hint = random.choice([loc for loc in slam_locations if loc not in hinted_slam_locations])
+            item_region_locations_to_hint.append(loc_to_hint)
+            optional_hintable_locations.remove(loc_to_hint)
         # Fill with other random move locations as best as we can
         random.shuffle(optional_hintable_locations)
         while len(item_region_locations_to_hint) < hint_distribution[HintType.ItemRegion] and len(optional_hintable_locations) > 0:
