+++ templates/plandomizer/plandomizer.html.jinja2
@@ -27,6 +27,7 @@
         <input id="import_plando_settings"
                class="btn btn-primary settings-button"
                type="button"
+               onClick="plando_import_filebox()"
                value="Import Settings from File"/>
         <input id="export_plando_settings"
                class="btn btn-primary settings-button"
